mesa: Tuning builds for various arches
When building for i386 on an x86_64 host, config.guess will assume
that it's an x86_64 build and this will result in x86_64 assembly
code features being used with a 32bit x86 assember (at least with
intel drivers enabled).

Explicitly telling the build scripts to build for i586 fixes this
(we give it i586 in the triple because that's the compiler we build
in the base runtime).