llvm: Set LLVM_HOST_TRIPLE when building for i386 or arm
When configuring llvm it will detect the host arch as the build
arch even when the host arch is a 64bit arch and the whole sysroot
you are building in is 32bit (i.e. when building i686 binaries
on x86_64 or when building armv7a binaries on an aarch64 host).

This patch forces the generated llvm to use the 32bit target
arch as it's default - in turn this causes clang to successfully
compile in the target 32bit sysroot.